refactor: extract record search components
